# Backend/openapi.yaml
openapi: 3.0.0
info:
  title: ML Model Serving Platform
  description: |
    # ML Model Serving Platform API

    A production-ready platform for deploying, managing, and serving machine learning models via REST API.

    ## Key Features

    - **Model Management**: Upload, version, and manage ML models
    - **Authentication**: JWT + API Key authentication with refresh tokens
    - **Predictions**: Real-time predictions with caching and performance tracking
    - **Sharing**: Share models with other users with granular permissions
    - **Analytics**: Track model performance and usage metrics
    - **Webhooks**: Event-driven architecture with webhook support
    - **Rate Limiting**: Per-user rate limiting with Redis backend
    - **Health Checks**: Kubernetes-ready health check endpoints

    ## Getting Started

    1. **Register**: `POST /api/v1/auth/register`
    2. **Login**: `POST /api/v1/auth/login`
    3. **Upload Model**: `POST /api/v1/models/upload`
    4. **Make Prediction**: `POST /api/v1/predictions/predict`

  version: "1.0.0"
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
servers:
  - url: "https://api.example.com"
    description: Production server
  - url: "http://localhost:8000"
    description: Local development server
...
paths:
...
  /api/v1/models/upload:
    post:
      summary: Upload a new ML model
      tags:
        - Models
      security:
        - BearerToken: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - model_file
                - model_name
              properties:
                model_file:
                  type: string
                  format: binary
                model_name:
                  type: string
                model_type:
                  type: string
                  enum:
                    - sklearn
                    - tensorflow
                    - pytorch
                description:
                  type: string
                version_note:
                  type: string
      responses:
        "201":
          description: Model uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  model_id:
                    type: string
                  model_name:
                    type: string
                  version:
                    type: integer
                  model_type:
                    type: string
                  status:
                    type: string
                  created_at:
                    type: string
                    format: date-time
  /api/v1/models:
    get:
      summary: List user's models
      tags:
        - Models
      security:
        - BearerToken: []
      parameters:
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: status
          in: query
          schema:
            type: string
            enum:
              - active
              - deprecated
              - archived
      responses:
        "200":
          description: List of models
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  total:
                    type: integer
                  skip:
                    type: integer
                  limit:
                    type: integer
  /api/v1/models/{model_id}:
    get:
      summary: Get model details
      tags:
        - Models
      security:
        - BearerToken: []
      parameters:
        - name: model_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Model details
          content:
            application/json:
              schema:
                type: object
                properties:
                  model_id:
                    type: string
                  model_name:
                    type: string
                  version:
                    type: integer
                  status:
                    type: string

    patch:
      summary: Update model metadata
      tags:
        - Models
      security:
        - BearerToken: []
      parameters:
        - name: model_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                status:
                  type: string
                metadata:
                  type: object
      responses:
        "200":
          description: Model updated

    delete:
      summary: Delete model
      tags:
        - Models
      security:
        - BearerToken: []
      parameters:
        - name: model_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Model deleted
...
components:
  securitySchemes:
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /auth/login

    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key from /api-keys/create
...